Move path concerns to path.py and template concerns to editor.py -- BDD tests are failing
