+++ jrnl/path.py
@@ -2,6 +2,19 @@
 # License: https://www.gnu.org/licenses/gpl-3.0.html
 
 import os.path
+from pathlib import Path
+
+import xdg.BaseDirectory
+
+from jrnl.config import DEFAULT_CONFIG_NAME
+from jrnl.config import DEFAULT_JOURNAL_NAME
+from jrnl.exception import JrnlException
+from jrnl.messages import Message
+from jrnl.messages import MsgStyle
+from jrnl.messages import MsgText
+
+# Constants
+XDG_RESOURCE = "jrnl"
 
 
 def home_dir() -> str:
@@ -14,3 +27,42 @@ def expand_path(path: str) -> str:
 
 def absolute_path(path: str) -> str:
     return os.path.abspath(expand_path(path))
+
+
+def get_default_journal_path() -> str:
+    journal_data_path = xdg.BaseDirectory.save_data_path(XDG_RESOURCE) or home_dir()
+    return os.path.join(journal_data_path, DEFAULT_JOURNAL_NAME)
+
+
+def get_templates_path() -> Path:
+    # jrnl_xdg_resource_path is created by save_data_path if it does not exist
+    jrnl_xdg_resource_path = Path(xdg.BaseDirectory.save_data_path(XDG_RESOURCE))
+    jrnl_templates_path = jrnl_xdg_resource_path / "templates"
+    # Create the directory if needed.
+    jrnl_templates_path.mkdir(exist_ok=True)
+    return jrnl_templates_path
+
+
+def get_config_directory() -> str:
+    try:
+        return xdg.BaseDirectory.save_config_path(XDG_RESOURCE)
+    except FileExistsError:
+        raise JrnlException(
+            Message(
+                MsgText.ConfigDirectoryIsFile,
+                MsgStyle.ERROR,
+                {
+                    "config_directory_path": os.path.join(
+                        xdg.BaseDirectory.xdg_config_home, XDG_RESOURCE
+                    )
+                },
+            ),
+        )
+
+
+def get_config_path() -> Path:
+    try:
+        config_directory_path = get_config_directory()
+    except JrnlException:
+        return Path(home_dir(), DEFAULT_CONFIG_NAME)
+    return Path(config_directory_path, DEFAULT_CONFIG_NAME)
